CRUD_ORION/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
requests
websocket-client